Fix sql_check example: cast attnum to integer and add colored output
- Fix ordinal column decode error by casting attnum::integer (smallint to i32)
- Add colored crate for pretty terminal output with colors
- Show green checkmark for valid SQL, red X for errors
- Display error details with level, kind, location, and caret indicator

+++ crates/pgorm-check/src/schema_introspect.rs
@@ -116,7 +116,7 @@ SELECT
   c.relname AS table_name,
   c.relkind AS relkind,
   a.attname AS column_name,
-  a.attnum AS ordinal,
+  a.attnum::integer AS ordinal,
   pg_catalog.format_type(a.atttypid, a.atttypmod) AS data_type,
   a.attnotnull AS not_null,
   pg_get_expr(ad.adbin, ad.adrelid) AS default_expr
